change op tensorType to LongTensor
Float tensor doesn't work with CrossEntropy loss

+++ Model/NQDataset.py
@@ -28,7 +28,7 @@ class NQDataset(Dataset) :
 		# the outputs expected
 		answer_type = self.get_ans_type(item)
 		short_start, short_end = self.get_short_start_end(item.short_ans, inputids.tolist()[0])
-		yes_no = torch.FloatTensor([item.yes_no_ans or 0])
+		yes_no = torch.LongTensor([item.yes_no_ans or 0])
 
 		return inputids, token_type, mask, answer_type, short_start, short_end, yes_no
 
@@ -45,7 +45,7 @@ class NQDataset(Dataset) :
 		elif item.yes_no_ans is not None : ans_type[2] = 1
 		elif item.short_ans is not None : ans_type[1]  = 1
 		else : error("get_ans_type in NQDataset", "no answer type found")
-		return torch.FloatTensor([ans_type])
+		return torch.LongTensor([ans_type])
 
 	def get_short_start_end(self, short_ans_list: List[str], long_list: List[int]) : 
 		'''
@@ -62,7 +62,7 @@ class NQDataset(Dataset) :
 			start_list[start] = 1
 			end_list[end] = 1
 
-		return torch.FloatTensor([start_list]), torch.FloatTensor([end_list])
+		return torch.LongTensor([start_list]), torch.LongTensor([end_list])
 
 
 	def get_span(self, short_string: str, long_list: List[int]) :
